chore: 패키지 구조 변경 - api 패키지 삭제 및 controller 상위 폴더로 이동
