--- src/main/java/com/hh99/ecommerce/balance/api/response/ChargeBalanceResponse.java
@@ -1,13 +1,0 @@
-package com.hh99.ecommerce.balance.api.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ChargeBalanceResponse {
-    private String message;
-    private int newBalance;
-}
