+++ src/main/java/com/hh99/ecommerce/balance/application/service/BalanceService.java
@@ -1,9 +1,13 @@
 package com.hh99.ecommerce.balance.application.service;
 
-import com.hh99.ecommerce.balance.api.response.BalanceResponse;
+import com.hh99.ecommerce.balance.controller.response.BalanceResponse;
 import com.hh99.ecommerce.balance.application.exception.InvalidAmountException;
 import com.hh99.ecommerce.balance.application.exception.UserBalanceNotFoundException;
-import com.hh99.ecommerce.balance.domain.*;
+import com.hh99.ecommerce.balance.domain.entity.Balance;
+import com.hh99.ecommerce.balance.domain.entity.BalanceHistory;
+import com.hh99.ecommerce.balance.domain.entity.BalanceHistoryType;
+import com.hh99.ecommerce.balance.domain.repository.BalanceHistoryRepository;
+import com.hh99.ecommerce.balance.domain.repository.BalanceRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
