+++ src/main/java/com/hh99/ecommerce/balance/controller/request/ChargeBalanceRequest.java
@@ -1,0 +1,16 @@
+package com.hh99.ecommerce.balance.controller.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ChargeBalanceRequest {
+    private Long userId;
+    private BigDecimal amount;
+}
+
