+++ src/main/java/com/hh99/ecommerce/balance/controller/response/BalanceResponse.java
@@ -1,0 +1,15 @@
+package com.hh99.ecommerce.balance.controller.response;
+
+import lombok.*;
+
+import java.math.BigDecimal;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class BalanceResponse {
+    private Long id;
+    private Long userId;
+    private BigDecimal amount;
+}
