+++ src/main/java/com/hh99/ecommerce/balance/controller/response/ChargeBalanceResponse.java
@@ -1,0 +1,13 @@
+package com.hh99.ecommerce.balance.controller.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ChargeBalanceResponse {
+    private String message;
+    private int newBalance;
+}
