--- src/main/java/com/hh99/ecommerce/balance/domain/BalanceHistoryRepository.java
@@ -1,8 +1,0 @@
-package com.hh99.ecommerce.balance.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BalanceHistoryRepository extends JpaRepository<BalanceHistory, Long> {
-}
