--- src/main/java/com/hh99/ecommerce/balance/domain/BalanceHistoryType.java
@@ -1,5 +1,0 @@
-package com.hh99.ecommerce.balance.domain;
-
-public enum BalanceHistoryType {
-    CHARGE, DEDUCT
-}
