--- src/main/java/com/hh99/ecommerce/balance/domain/BalanceRepository.java
@@ -1,12 +1,0 @@
-package com.hh99.ecommerce.balance.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface BalanceRepository extends JpaRepository<Balance, Long> {
-
-    Optional<Balance> findByUserId(Long userId);
-}
