+++ src/main/java/com/hh99/ecommerce/balance/domain/entity/Balance.java
@@ -1,0 +1,32 @@
+package com.hh99.ecommerce.balance.domain.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
+@Entity
+@Table(name = "balance")
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Balance {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "user_id", nullable = false)
+    private Long userId;
+
+    @Column(nullable = false)
+    private BigDecimal amount;
+
+    public void charge(BigDecimal chargeAmount) {
+        this.amount = this.amount.add(chargeAmount);
+    }
+}
