+++ src/main/java/com/hh99/ecommerce/balance/domain/entity/BalanceHistoryType.java
@@ -1,0 +1,5 @@
+package com.hh99.ecommerce.balance.domain.entity;
+
+public enum BalanceHistoryType {
+    CHARGE, DEDUCT
+}
