+++ src/main/java/com/hh99/ecommerce/balance/domain/repository/BalanceRepository.java
@@ -1,0 +1,13 @@
+package com.hh99.ecommerce.balance.domain.repository;
+
+import com.hh99.ecommerce.balance.domain.entity.Balance;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface BalanceRepository extends JpaRepository<Balance, Long> {
+
+    Optional<Balance> findByUserId(Long userId);
+}
