--- src/main/java/com/hh99/ecommerce/order/api/request/CreateOrderItemRequest.java
@@ -1,15 +1,0 @@
-package com.hh99.ecommerce.order.api.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class CreateOrderItemRequest {
-    private Long productId;
-    private int quantity;
-}
