--- src/main/java/com/hh99/ecommerce/order/api/request/CreateOrderRequest.java
@@ -1,18 +1,0 @@
-package com.hh99.ecommerce.order.api.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class CreateOrderRequest {
-    private Long userId;
-    private Long productId;
-    private int quantity;
-}
-
