--- src/main/java/com/hh99/ecommerce/order/api/response/CreateOrderResponse.java
@@ -1,13 +1,0 @@
-package com.hh99.ecommerce.order.api.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class CreateOrderResponse {
-    private String message;
-    private Long orderId;
-}
