--- src/main/java/com/hh99/ecommerce/order/api/response/OrderItemResponse.java
@@ -1,18 +1,0 @@
-package com.hh99.ecommerce.order.api.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class OrderItemResponse {
-    private Long id;
-    private Long orderId;
-    private Long productId;
-    private Integer quantity;
-    private BigDecimal itemPrice;
-}
