--- src/main/java/com/hh99/ecommerce/order/api/response/OrderResponse.java
@@ -1,17 +1,0 @@
-package com.hh99.ecommerce.order.api.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class OrderResponse {
-   private Long id;
-   private LocalDateTime orderDate;
-   private BigDecimal totalPrice;
-}
