+++ src/main/java/com/hh99/ecommerce/order/controller/response/CreateOrderResponse.java
@@ -1,0 +1,13 @@
+package com.hh99.ecommerce.order.controller.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class CreateOrderResponse {
+    private String message;
+    private Long orderId;
+}
