+++ src/main/java/com/hh99/ecommerce/order/domain/entity/Order.java
@@ -1,0 +1,33 @@
+package com.hh99.ecommerce.order.domain.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "`order`")
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class Order {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "user_id", nullable = false)
+    private Long userId;
+
+    @Column(name = "order_date", nullable = false)
+    private LocalDateTime orderDate;
+
+    @Column(name = "total_price", nullable = false)
+    private BigDecimal totalPrice;
+
+}
