+++ src/main/java/com/hh99/ecommerce/order/domain/repository/OrderItemRepository.java
@@ -1,6 +1,6 @@
 package com.hh99.ecommerce.order.domain.repository;
 
-import com.hh99.ecommerce.order.domain.OrderItem;
+import com.hh99.ecommerce.order.domain.entity.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
