+++ src/main/java/com/hh99/ecommerce/order/domain/repository/OrderRepository.java
@@ -1,6 +1,6 @@
 package com.hh99.ecommerce.order.domain.repository;
 
-import com.hh99.ecommerce.order.domain.Order;
+import com.hh99.ecommerce.order.domain.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
