--- src/main/java/com/hh99/ecommerce/product/api/request/ProductRequest.java
@@ -1,18 +1,0 @@
-package com.hh99.ecommerce.product.api.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ProductRequest {
-    private Long id;
-    private String name;
-    private String description;
-    private BigDecimal price;
-    private int stock;
-}
