--- src/main/java/com/hh99/ecommerce/product/api/response/ProductResponse.java
@@ -1,22 +1,0 @@
-package com.hh99.ecommerce.product.api.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ProductResponse {
-    private Long id;
-    private String name;
-    private String description;
-    private BigDecimal price;
-    private int stock;
-    private LocalDateTime regDate;
-}
