+++ src/main/java/com/hh99/ecommerce/product/controller/response/ProductResponse.java
@@ -1,0 +1,22 @@
+package com.hh99.ecommerce.product.controller.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ProductResponse {
+    private Long id;
+    private String name;
+    private String description;
+    private BigDecimal price;
+    private int stock;
+    private LocalDateTime regDate;
+}
